requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
Django==3.1.2
django-bootstrap4==2.2.0
django-environ==0.4.5
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.4.1
